fix issue #4 by changing dtype of action_indices

+++ ReinforcementLearning/DeepQLearning/simple_dqn_torch.py
@@ -80,7 +80,7 @@ class Agent(object):
             batch = np.random.choice(max_mem, self.batch_size)
             state_batch = self.state_memory[batch]
             action_batch = self.action_memory[batch]
-            action_values = np.array(self.action_space, dtype=np.uint8)
+            action_values = np.array(self.action_space, dtype=np.int32)
             action_indices = np.dot(action_batch, action_values)
             reward_batch = self.reward_memory[batch]
             new_state_batch = self.new_state_memory[batch]
